requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
aiofiles
youtube-transcript-api
supabase
pyjwt[crypto]
cryptography
